[compiler] fix DeviceGraphCluster for if op
