+++ compiler/include/byteir/Analysis/ShapeAnalysis.h
@@ -20,28 +20,36 @@
 
 #include "mlir/Analysis/DataFlow/ConstantPropagationAnalysis.h"
 #include "mlir/Analysis/DataFlow/SparseAnalysis.h"
+#include "mlir/Dialect/Arith/IR/Arith.h"
+#include "mlir/Dialect/Shape/IR/Shape.h"
+#include "mlir/Dialect/Tensor/IR/Tensor.h"
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/Types.h"
 #include "mlir/Interfaces/InferTypeOpInterface.h"
 #include "llvm/ADT/Sequence.h"
 #include "llvm/ADT/SmallVector.h"
+#include "llvm/ADT/TypeSwitch.h"
+#include "llvm/Support/Debug.h"
+
+#define DEBUG_TYPE "shape-analysis"
 
 namespace mlir {
 namespace shape_analysis {
 /// Statically known information for a particular Value.
 ///
 /// This struct currently tracks only information relevant for tensor/array-like
-/// shaped types. It is fine to associate a `ValueKnowledge` with a non-shaped
-/// type as long as it is in the default "no knowledge" state returned by
-/// `getPessimisticValueState`. The important invariant is that we cannot
-/// claim to know something about a value which is false.
+/// shaped types. It is fine to associate a `StaticShapeKnowledge` with a
+/// non-shaped type as long as it is in the default "no knowledge" state
+/// returned by `getPessimisticValueState`. The important invariant is that we
+/// cannot claim to know something about a value which is false.
 ///
 /// This class could also be called "dataflow facts", "lattice value", etc.
-struct ValueKnowledge {
-  ValueKnowledge() : hasError(false), hasRank(false), dtype(std::nullopt) {}
+struct StaticShapeKnowledge {
+  StaticShapeKnowledge()
+      : hasError(false), hasRank(false), dtype(std::nullopt) {}
 
-  ValueKnowledge(bool hasRank, llvm::ArrayRef<int64_t> newSizes,
-                 std::optional<Type> dtype)
+  StaticShapeKnowledge(bool hasRank, llvm::ArrayRef<int64_t> newSizes,
+                       std::optional<Type> dtype)
       : hasError(false), hasRank(hasRank), dtype(dtype) {
     sizes.reserve(newSizes.size());
     for (auto size : newSizes)
@@ -51,13 +59,13 @@ struct ValueKnowledge {
   operator bool() const { return !hasError; }
 
   // Get the static knowledge intrinsic to `type`.
-  static ValueKnowledge getKnowledgeFromType(Type type);
+  static StaticShapeKnowledge getKnowledgeFromType(Type type);
 
   // Return a pessimistic/conservative value state without assuming any knowlege
   // about the IR.
-  static ValueKnowledge getPessimisticValueState();
+  static StaticShapeKnowledge getPessimisticValueState();
 
-  static ValueKnowledge getPessimisticValueState(Value value);
+  static StaticShapeKnowledge getPessimisticValueState(Value value);
 
   /// Whether the state is uninitialized.
   bool isUninitialized() const { return !dtype.has_value(); }
@@ -72,15 +80,15 @@ struct ValueKnowledge {
     return UnrankedTensorType::get(*dtype);
   }
 
-  bool operator==(const ValueKnowledge &rhs) const {
+  bool operator==(const StaticShapeKnowledge &rhs) const {
     return hasRank == rhs.hasRank && sizes == rhs.sizes && dtype == rhs.dtype;
   }
 
-  static ValueKnowledge join(const ValueKnowledge &lhs,
-                             const ValueKnowledge &rhs);
+  static StaticShapeKnowledge join(const StaticShapeKnowledge &lhs,
+                                   const StaticShapeKnowledge &rhs);
 
-  static ValueKnowledge meet(const ValueKnowledge &lhs,
-                             const ValueKnowledge &rhs);
+  static StaticShapeKnowledge meet(const StaticShapeKnowledge &lhs,
+                                   const StaticShapeKnowledge &rhs);
 
   void print(raw_ostream &os) const;
 
@@ -93,8 +101,25 @@ struct ValueKnowledge {
   llvm::SmallVector<int64_t> sizes;
   // The dtype of a tensor.
   // This is equal to nullptr if we don't know that it is a specific concrete
-  // type. The ValueKnowledge object is isUninitialized if dtype is None.
+  // type. The StaticShapeKnowledge object is isUninitialized if dtype is None.
   std::optional<Type> dtype;
+};
+
+class BoundedShapeKnowledge : public StaticShapeKnowledge {
+public:
+  BoundedShapeKnowledge() : StaticShapeKnowledge() {}
+
+  BoundedShapeKnowledge(bool hasRank, llvm::ArrayRef<int64_t> newSizes,
+                        std::optional<Type> dtype)
+      : StaticShapeKnowledge(hasRank, newSizes, dtype) {}
+
+  static BoundedShapeKnowledge getKnowledgeFromType(Type type);
+  static BoundedShapeKnowledge getPessimisticValueState();
+  static BoundedShapeKnowledge getPessimisticValueState(Value value);
+  static BoundedShapeKnowledge join(const BoundedShapeKnowledge &lhs,
+                                    const BoundedShapeKnowledge &rhs);
+  static BoundedShapeKnowledge meet(const BoundedShapeKnowledge &lhs,
+                                    const BoundedShapeKnowledge &rhs);
 };
 
 struct ValueTypeModificatoinRAII {
@@ -113,102 +138,271 @@ struct ValueTypeModificatoinRAII {
   SmallVector<std::pair<Value, Type>> toRestore;
 };
 } // namespace shape_analysis
-
-namespace value_analysis {
-struct BoundedValueKnowledge {
-  struct BoundedValue {
-    Attribute lower;
-    Attribute upper;
-    bool operator==(const BoundedValue &rhs) const {
-      return lower == rhs.lower && upper == rhs.upper;
-    }
-    bool isUnknwon() const { return (!lower) || (!upper); }
-  };
-  explicit BoundedValueKnowledge() = default;
-  explicit BoundedValueKnowledge(BoundedValue bdValue)
-      : boundedValue(bdValue) {}
-  explicit BoundedValueKnowledge(Attribute lower, Attribute upper) {
-    BoundedValue bv;
-    bv.lower = lower;
-    bv.upper = upper;
-    boundedValue = bv;
-  }
-
-  // Get the static knowledge intrinsic to `type`.
-  static BoundedValueKnowledge getUninitializedValue();
-
-  static BoundedValueKnowledge getUnknownValue();
-
-  static BoundedValueKnowledge getKnownValue(Attribute lower, Attribute upper);
-
-  static BoundedValueKnowledge join(const BoundedValueKnowledge &lhs,
-                                    const BoundedValueKnowledge &rhs);
-
-  static BoundedValueKnowledge meet(const BoundedValueKnowledge &lhs,
-                                    const BoundedValueKnowledge &rhs);
-
-  /// Whether the state is uninitialized.
-  bool isUninitialized() const { return !boundedValue.has_value(); }
-
-  bool isUnknown() const {
-    return boundedValue.has_value() && boundedValue->isUnknwon();
-  }
-
-  bool isKnown() const {
-    return boundedValue.has_value() && !boundedValue->isUnknwon();
-  }
-
-  bool operator==(const BoundedValueKnowledge &rhs) const {
-    return boundedValue == rhs.boundedValue;
-  }
-
-  Attribute lower() const;
-  Attribute upper() const;
-
-  void print(raw_ostream &os) const;
-
-  std::optional<BoundedValue> boundedValue;
-};
-} // namespace value_analysis
-
-using ShapeLattice = dataflow::Lattice<shape_analysis::ValueKnowledge>;
-
-class ShapeAnalysis
-    : public dataflow::SparseForwardDataFlowAnalysis<ShapeLattice> {
-public:
-  using SparseForwardDataFlowAnalysis::SparseForwardDataFlowAnalysis;
-
-  void visitOperation(Operation *op, ArrayRef<const ShapeLattice *> operands,
-                      ArrayRef<ShapeLattice *> results) override;
-
-  void setToEntryState(ShapeLattice *lattice) override;
-
-protected:
-  using ShapeKnowledges = function_ref<Type(Value)>;
-  using ShapeValueKnowledges = function_ref<Attribute(Value)>;
-
-  virtual LogicalResult inferResultShapesWithKnowledges(
-      Operation *op, ShapeKnowledges shapeKnowledges,
-      ShapeValueKnowledges shapeValueKnowledges,
-      llvm::SmallVectorImpl<::mlir::ShapedTypeComponents> &results);
-};
-
 // FIXME: strictly speaking ShapeValueLattice should be a 1-d tensor which could
 // be inferred partially, here we use the same Lattice as the CPA did, so once
 // the state of the Lattice is mutated the subscribed CPA would be triggered
 using ShapeValueLattice = dataflow::Lattice<dataflow::ConstantValue>;
 
+using ShapeKnowledges = function_ref<Type(Value)>;
+using ShapeValueKnowledges = function_ref<Attribute(Value)>;
+
+template <typename ShapeKnowledgeType>
+class ShapeAnalysis : public dataflow::SparseForwardDataFlowAnalysis<
+                          dataflow::Lattice<ShapeKnowledgeType>> {
+public:
+  using ShapeLattice = dataflow::Lattice<ShapeKnowledgeType>;
+  using BaseT = dataflow::SparseForwardDataFlowAnalysis<ShapeLattice>;
+  using BaseT::SparseForwardDataFlowAnalysis;
+
+  void visitOperation(Operation *op, ArrayRef<const ShapeLattice *> operands,
+                      ArrayRef<ShapeLattice *> results) override {
+
+    LLVM_DEBUG(llvm::dbgs() << "shape analysis on " << *op << "\n");
+
+    llvm::DenseMap<Value, Type> shapeProvider;
+    llvm::DenseMap<Value, Attribute> valueProvider;
+    bool missingValue = false;
+    for (auto &&pi : llvm::zip(op->getOperands(), operands)) {
+      auto &&operand = std::get<0>(pi);
+      auto &&shapeLattice = std::get<1>(pi);
+      auto &&valueLattice =
+          this->template getOrCreate<ShapeValueLattice>(operand);
+      valueLattice->useDefSubscribe(this);
+
+      if (auto shapeKnowledge = shapeLattice->getValue()) {
+        if (shapeKnowledge.isUninitialized()) {
+          missingValue = true;
+          continue;
+        }
+      }
+
+      if (valueLattice->getValue().isUninitialized()) {
+        missingValue = true;
+        continue;
+      }
+
+      if (auto shapeKnowledge = shapeLattice->getValue()) {
+        if (*shapeKnowledge.dtype) {
+          shapeProvider[operand] = shapeKnowledge.getType();
+        }
+      }
+
+      if (valueLattice->getValue().getConstantValue()) {
+        valueProvider[operand] = valueLattice->getValue().getConstantValue();
+      }
+    }
+
+    if (missingValue) {
+      return;
+    }
+
+    auto shapeKnowledges = [&](Value val) -> Type {
+      auto it = shapeProvider.find(val);
+      if (it == shapeProvider.end())
+        return nullptr;
+      return it->second;
+    };
+    auto shapeValueKnowledges = [&](Value val) -> Attribute {
+      auto it = valueProvider.find(val);
+      if (it == valueProvider.end())
+        return nullptr;
+      return it->second;
+    };
+
+    SmallVector<ShapedTypeComponents> inferredShapes;
+    if (inferResultShapesWithKnowledges(op, shapeKnowledges,
+                                        shapeValueKnowledges, inferredShapes)
+            .succeeded()) {
+      for (auto it : llvm::zip(op->getResults(), inferredShapes, results)) {
+        Value result = std::get<0>(it);
+        ShapedTypeComponents predictedShape = std::get<1>(it);
+        ShapeLattice *resultLattice = std::get<2>(it);
+
+        Type resultTy = result.getType();
+        if (!isa<ShapedType>(resultTy)) {
+          setToEntryState(resultLattice);
+          continue;
+        }
+
+        // Compute the knowledge based on the inferred type.
+        auto inferredKnowledge = ShapeKnowledgeType::getPessimisticValueState();
+        inferredKnowledge.dtype = cast<ShapedType>(resultTy).getElementType();
+        inferredKnowledge.hasRank = predictedShape.hasRank();
+        if (predictedShape.hasRank()) {
+          for (auto dim : predictedShape.getDims()) {
+            inferredKnowledge.sizes.push_back(dim);
+          }
+        }
+
+        this->template propagateIfChanged(
+            resultLattice, resultLattice->join(inferredKnowledge));
+      }
+    } else {
+      return this->template setAllToEntryStates(results);
+    }
+  }
+
+  void setToEntryState(ShapeLattice *lattice) override {
+    this->template propagateIfChanged(
+        lattice, lattice->join(ShapeKnowledgeType::getPessimisticValueState(
+                     lattice->getPoint())));
+  }
+
+protected:
+  virtual LogicalResult inferResultShapesWithKnowledges(
+      Operation *op, ShapeKnowledges shapeKnowledges,
+      ShapeValueKnowledges shapeValueKnowledges,
+      llvm::SmallVectorImpl<::mlir::ShapedTypeComponents> &results) {
+    if (op->hasTrait<OpTrait::SameOperandsAndResultShape>()) {
+      ShapeKnowledgeType knowledge; // uninitialized
+      for (auto &&operand : op->getOperands()) {
+        auto newKnowledge =
+            ShapeKnowledgeType::getKnowledgeFromType(shapeKnowledges(operand));
+        newKnowledge.dtype = nullptr;
+        knowledge = ShapeKnowledgeType::meet(knowledge, newKnowledge);
+      }
+      if (knowledge) {
+        for (auto &&resultType : op->getResultTypes()) {
+          if (auto shapedType = dyn_cast_or_null<ShapedType>(resultType)) {
+            knowledge.dtype = shapedType.getElementType();
+            results.push_back(cast<ShapedType>(knowledge.getType()));
+          } else {
+            results.push_back(ShapedTypeComponents{});
+          }
+        }
+        return success();
+      }
+    }
+
+    if (auto shapeInterface = dyn_cast<InferShapedTypeOpInterface>(op)) {
+      shape_analysis::ValueTypeModificatoinRAII valueTypeModification;
+      for (auto &&operand : op->getOperands()) {
+        if (auto shape = shapeKnowledges(operand)) {
+          valueTypeModification.Push(operand, shape);
+        }
+      }
+      ValueShapeRange range(op->getOperands(), shapeKnowledges,
+                            shapeValueKnowledges);
+      if (shapeInterface
+              .inferReturnTypeComponents(op->getContext(), op->getLoc(), range,
+                                         op->getAttrDictionary(),
+                                         op->getPropertiesStorage(),
+                                         op->getRegions(), results)
+              .succeeded()) {
+        return success();
+      }
+    }
+
+    if (auto typeInterface = dyn_cast<InferTypeOpInterface>(op)) {
+      shape_analysis::ValueTypeModificatoinRAII valueTypeModification;
+      for (auto &&operand : op->getOperands()) {
+        if (auto shape = shapeKnowledges(operand)) {
+          valueTypeModification.Push(operand, shape);
+        }
+      }
+      llvm::SmallVector<Type> inferredType;
+      if (typeInterface
+              .inferReturnTypes(op->getContext(), op->getLoc(),
+                                op->getOperands(), op->getAttrDictionary(),
+                                op->getPropertiesStorage(), op->getRegions(),
+                                inferredType)
+              .succeeded()) {
+        results.assign(llvm::to_vector(llvm::map_range(
+            inferredType, [](mlir::Type t) -> ShapedTypeComponents {
+              if (auto st = dyn_cast_or_null<ShapedType>(t))
+                return st;
+              return {};
+            })));
+        return success();
+      }
+    }
+
+    return failure();
+  }
+};
+
 // derived from SCP but override for some operation which could be folded with
 // operand type instead of operand value, for such operation it should not mark
 // them as pessimistic fixpoint when fold failed with given operand value, it
 // might be updated once operand type is inferred
+template <typename ShapeKnowledgeType>
 class ShapeValueAnalysis : public dataflow::SparseConstantPropagation {
 public:
-  using SparseConstantPropagation::SparseConstantPropagation;
+  using dataflow::SparseConstantPropagation::SparseConstantPropagation;
+  using ShapeLattice = dataflow::Lattice<ShapeKnowledgeType>;
 
   void visitOperation(Operation *op,
                       ArrayRef<const ShapeValueLattice *> operands,
-                      ArrayRef<ShapeValueLattice *> results) override;
+                      ArrayRef<ShapeValueLattice *> results) override {
+    LLVM_DEBUG(llvm::dbgs() << "shape value analysis on " << *op << "\n");
+    TypeSwitch<Operation *>(op)
+        .template Case<shape::ShapeOfOp>([&](Operation *op) {
+          auto *shapeLattice = getOrCreate<ShapeLattice>(op->getOperand(0));
+          shapeLattice->useDefSubscribe(this);
+          if (shapeLattice->getValue().isUninitialized()) {
+            return;
+          }
+          auto inputType =
+              dyn_cast<RankedTensorType>(shapeLattice->getValue().getType());
+          if (!inputType || !inputType.hasStaticShape()) {
+            return setAllToEntryStates(results);
+          }
+          auto shape = inputType.getShape();
+          auto outType = dyn_cast<RankedTensorType>(op->getResult(0).getType());
+          auto resultAttr = DenseIntElementsAttr::get(outType, shape);
+          auto lattice = results[0];
+          propagateIfChanged(lattice, lattice->join(dataflow::ConstantValue(
+                                          resultAttr, op->getDialect())));
+        })
+        .template Case<tensor::DimOp>([&](Operation *op) {
+          SmallVector<const ShapeLattice *> shapeLattices(op->getNumOperands(),
+                                                          nullptr);
+          shapeLattices[0] = getOrCreate<ShapeLattice>(op->getOperand(0));
+          visitOperation(op, operands, shapeLattices, results);
+        })
+        .template Case<arith::IndexCastOp>([&](Operation *op) {
+          const ShapeValueLattice *index = operands[0];
+          if (index->getValue().isUninitialized()) {
+            return;
+          }
+          Attribute constAttr = index->getValue().getConstantValue();
+          if (auto denseInt =
+                  dyn_cast_or_null<DenseIntElementsAttr>(constAttr)) {
+
+            auto newType = denseInt.getType().clone(
+                cast<RankedTensorType>(cast<arith::IndexCastOp>(op).getType())
+                    .getElementType());
+
+            SmallVector<APInt> newDenseInt;
+            uint32_t width;
+            auto elemType = newType.getElementType();
+            if (elemType.isIntOrFloat()) {
+              width = elemType.getIntOrFloatBitWidth();
+            } else {
+              assert(isa<IndexType>(elemType));
+              width = IndexType::kInternalStorageBitWidth;
+            }
+
+            for (auto i : denseInt.getValues<APInt>()) {
+              newDenseInt.push_back(APInt(width, i.getZExtValue()));
+            }
+
+            auto resultAttr = DenseElementsAttr::get(newType, newDenseInt);
+
+            auto lattice = results[0];
+            propagateIfChanged(lattice, lattice->join(dataflow::ConstantValue(
+                                            resultAttr, op->getDialect())));
+
+          } else {
+            dataflow::SparseConstantPropagation::visitOperation(op, operands,
+                                                                results);
+          }
+        })
+        .template Default([&](Operation *op) {
+          dataflow::SparseConstantPropagation::visitOperation(op, operands,
+                                                              results);
+        });
+  }
 
 protected:
   // very similar to SparseConstantPropagation but fold \p op with given
@@ -216,20 +410,34 @@ protected:
   virtual void visitOperation(Operation *op,
                               ArrayRef<const ShapeValueLattice *> operands,
                               ArrayRef<const ShapeLattice *> ShapeLattices,
-                              ArrayRef<ShapeValueLattice *> results);
+                              ArrayRef<ShapeValueLattice *> results) {
+    shape_analysis::ValueTypeModificatoinRAII valueTypeModification;
+
+    bool missingShape = false;
+    for (auto &&pi : zip(op->getOperands(), ShapeLattices)) {
+      auto shapeLattice = std::get<1>(pi);
+      if (shapeLattice) {
+        const_cast<ShapeLattice *>(shapeLattice)->useDefSubscribe(this);
+
+        if (!shapeLattice->getValue().isUninitialized()) {
+          auto shapeKnowledge = shapeLattice->getValue();
+          if (shapeKnowledge && shapeKnowledge.dtype) {
+            valueTypeModification.Push(std::get<0>(pi),
+                                       shapeKnowledge.getType());
+            continue;
+          }
+        } else {
+          missingShape = true;
+        }
+      }
+    }
+    if (missingShape)
+      return;
+
+    dataflow::SparseConstantPropagation::visitOperation(op, operands, results);
+  }
 };
-
-using BoundedValueLattice =
-    dataflow::Lattice<value_analysis::BoundedValueKnowledge>;
-
-class BoundedValueAnalysis
-    : public dataflow::SparseForwardDataFlowAnalysis<BoundedValueLattice> {
-public:
-  using SparseForwardDataFlowAnalysis::SparseForwardDataFlowAnalysis;
-
-  void setToEntryState(BoundedValueLattice *lattice) override;
-};
-
 } // namespace mlir
+#undef DEBUG_TYPE
 
 #endif // BYTEIR_ANALYSIS_SHAPEANALYSIS_H
