+++ compiler/lib/Dialect/mhlo/Analysis/ShapeAnalysis.cpp
@@ -17,70 +17,102 @@
 
 #include "byteir/Dialect/mhlo/Analysis/ShapeAnalysis.h"
 #include "byteir/Dialect/mhlo/DynamicShapeOpRegister/Register.h"
-#include "byteir/Dialect/mhlo/Util/ShapeInferUtil.h"
-#include "byteir/Utils/Utils.h"
-#include "mhlo/IR/hlo_ops.h"
-#include "mlir/Dialect/Func/IR/FuncOps.h"
-#include "mlir/IR/BuiltinTypes.h"
-#include "llvm/ADT/StringMap.h"
-#include "llvm/ADT/TypeSwitch.h"
-#include "llvm/Support/Debug.h"
-#include <cassert>
 
 #define DEBUG_TYPE "mhlo-shape-analysis"
 
 using namespace mlir::shape_analysis;
 using namespace mlir::value_analysis;
 
-#define K_INITIAL -999
-
 namespace mlir {
-LogicalResult MhloShapeAnalysis::inferResultShapesWithKnowledges(
-    Operation *op, ShapeKnowledges shapeKnowledges,
-    ShapeValueKnowledges shapeValueKnowledges,
-    llvm::SmallVectorImpl<::mlir::ShapedTypeComponents> &results) {
-  InferReturnTypeComponents inferFunc = nullptr;
-  if (auto customCall = dyn_cast<mhlo::CustomCallOp>(op)) {
-    inferFunc = inferReturnTypeComponents(customCall.getCallTargetName());
-  } else {
-    inferFunc = inferReturnTypeComponents(op->getName().getStringRef());
-  }
-  if (nullptr == inferFunc) {
-    // fallback to generic shape analysis
-    return ShapeAnalysis::inferResultShapesWithKnowledges(
-        op, shapeKnowledges, shapeValueKnowledges, results);
-  }
-  ValueTypeModificatoinRAII valueTypeModification;
-  for (auto &&operand : op->getOperands()) {
-    Type newType = operand.getType();
-    if (auto shape = shapeKnowledges(operand)) {
-      newType = shape;
-    }
-    if (newType != operand.getType()) {
-      valueTypeModification.Push(operand, newType);
-    }
-  }
 
-  //  if return Attr{nullptr}, Type{nullptr} directly, ShapeAdaptor would try
-  //  dync_cast<> which cause crash
-  auto wrapperShapeKnowledges = [&](Value v) -> ShapeAdaptor {
-    if (auto type = shapeKnowledges(v)) {
-      return type;
-    }
-    return nullptr;
-  };
-  auto wrapperShapeValueKnowledges = [&](Value v) -> ShapeAdaptor {
-    if (auto attr = shapeValueKnowledges(v)) {
-      return attr;
-    }
-    return nullptr;
-  };
-  ValueShapeRange range(op->getOperands(), wrapperShapeKnowledges,
-                        wrapperShapeValueKnowledges);
-
-  return inferFunc(op->getContext(), op->getLoc(), range,
-                   op->getAttrDictionary(), op->getRegions(), results);
+namespace value_analysis {
+BoundedValueKnowledge BoundedValueKnowledge::getUninitializedValue() {
+  return BoundedValueKnowledge();
 }
+
+BoundedValueKnowledge BoundedValueKnowledge::getUnknownValue() {
+  BoundedValue boundedValue = {Attribute(), Attribute()};
+  BoundedValueKnowledge bv;
+  bv.boundedValue = boundedValue;
+  return bv;
+}
+BoundedValueKnowledge BoundedValueKnowledge::getKnownValue(Attribute lower,
+                                                           Attribute upper) {
+  assert(lower);
+  assert(upper);
+  BoundedValue boundedValue = {lower, upper};
+  BoundedValueKnowledge bv;
+  bv.boundedValue = boundedValue;
+  return bv;
+}
+
+Attribute BoundedValueKnowledge::lower() const { return (*boundedValue).lower; }
+
+Attribute BoundedValueKnowledge::upper() const { return (*boundedValue).upper; }
+
+BoundedValueKnowledge
+BoundedValueKnowledge::join(const BoundedValueKnowledge &lhs,
+                            const BoundedValueKnowledge &rhs) {
+  if (lhs.isUninitialized())
+    return rhs;
+  if (rhs.isUninitialized())
+    return lhs;
+  if (lhs == rhs)
+    return lhs;
+  return getUnknownValue();
+}
+
+BoundedValueKnowledge
+BoundedValueKnowledge::meet(const BoundedValueKnowledge &lhs,
+                            const BoundedValueKnowledge &rhs) {
+  auto res = getUnknownValue();
+  if (lhs.isUninitialized())
+    return rhs;
+  if (rhs.isUninitialized())
+    return lhs;
+
+  if (!((*(lhs.boundedValue)).lower)) {
+    (*(res.boundedValue)).lower = (*(rhs.boundedValue)).lower;
+  } else if (!((*(rhs.boundedValue)).lower)) {
+    (*(res.boundedValue)).lower = (*(lhs.boundedValue)).lower;
+  } else if ((*(lhs.boundedValue)).lower == (*(rhs.boundedValue)).lower) {
+    (*(res.boundedValue)).lower = (*(rhs.boundedValue)).lower;
+  } else {
+    (*(res.boundedValue)).lower = Attribute();
+  }
+
+  if (!((*(lhs.boundedValue)).upper)) {
+    (*(res.boundedValue)).upper = (*(rhs.boundedValue)).upper;
+  } else if (!((*(rhs.boundedValue)).upper)) {
+    (*(res.boundedValue)).upper = (*(lhs.boundedValue)).upper;
+  } else if ((*(lhs.boundedValue)).upper == (*(rhs.boundedValue)).upper) {
+    (*(res.boundedValue)).upper = (*(rhs.boundedValue)).upper;
+  } else {
+    (*(res.boundedValue)).upper = Attribute();
+  }
+  return res;
+}
+
+void BoundedValueKnowledge::print(raw_ostream &os) const {
+  if (isUninitialized()) {
+    os << "None\n";
+  } else if (isUnknown()) {
+    if ((*boundedValue).lower) {
+      os << "lower: " << (*boundedValue).lower << "\n";
+    } else {
+      os << "lower: Unknown\n";
+    }
+    if ((*boundedValue).upper) {
+      os << "upper: " << (*boundedValue).upper << "\n";
+    } else {
+      os << "upper: Unknown\n";
+    }
+  } else {
+    os << "lower: " << (*boundedValue).lower << "\n";
+    os << "upper: " << (*boundedValue).upper << "\n";
+  }
+}
+} // namespace value_analysis
 
 LogicalResult MhloBoundedShapeAnalysis::inferResultShapesWithKnowledges(
     Operation *op, ShapeKnowledges shapeKnowledges,
@@ -96,8 +128,9 @@ LogicalResult MhloBoundedShapeAnalysis::inferResultShapesWithKnowledges(
 
   if (nullptr == inferFunc) {
     // fallback to static mhlo shape analysis
-    return MhloShapeAnalysis::inferResultShapesWithKnowledges(
-        op, shapeKnowledges, shapeValueKnowledges, results);
+    return MhloShapeAnalysisBase<shape_analysis::BoundedShapeKnowledge>::
+        inferResultShapesWithKnowledges(op, shapeKnowledges,
+                                        shapeValueKnowledges, results);
   }
 
   ValueTypeModificatoinRAII valueTypeModification;
@@ -221,7 +254,8 @@ void MhloBoundedShapeAnalysis::visitOperation(
       }
 
       // Compute the knowledge based on the inferred type.
-      auto inferredKnowledge = ValueKnowledge::getPessimisticValueState();
+      auto inferredKnowledge =
+          shape_analysis::BoundedShapeKnowledge::getPessimisticValueState();
       inferredKnowledge.dtype = cast<ShapedType>(resultTy).getElementType();
       inferredKnowledge.hasRank = predictedShape.hasRank();
       if (predictedShape.hasRank()) {
@@ -235,147 +269,6 @@ void MhloBoundedShapeAnalysis::visitOperation(
   } else {
     return setAllToEntryStates(results);
   }
-}
-
-template <typename T>
-DenseElementsAttr PadOpFold(DenseElementsAttr input, T padValue,
-                            RankedTensorType returnType,
-                            DenseIntElementsAttr edgePaddingLow,
-                            DenseIntElementsAttr edgePaddingHigh,
-                            DenseIntElementsAttr interiorPadding) {
-  if (!input.getType().hasStaticShape() ||
-      !edgePaddingLow.getType().hasStaticShape() ||
-      !edgePaddingHigh.getType().hasStaticShape()) {
-    return {};
-  }
-  if (interiorPadding && !interiorPadding.getType().hasStaticShape())
-    return {};
-  // Fill the full result tensor with the padding value.
-  llvm::SmallVector<T> result(returnType.getNumElements(), padValue);
-
-  auto nextIndex = [](llvm::SmallVector<uint64_t> &index,
-                      llvm::ArrayRef<int64_t> shape) {
-    for (int64_t i = index.size() - 1; i >= 0; --i) {
-      ++index[i];
-      if (static_cast<int64_t>(index[i]) < shape[i])
-        return;
-      index[i] = 0;
-    }
-  };
-
-  // Iterate over all elements of the input tensor and copy it to the correct
-  // location in the output tensor.
-  llvm::SmallVector<uint64_t> index(input.getType().getRank(), 0);
-  uint64_t numElements = input.getNumElements();
-  for (uint64_t operandIdx = 0; operandIdx < numElements; operandIdx++) {
-    bool valid = true;
-    for (int i = 0; i < input.getType().getRank(); ++i) {
-      int64_t lowPad = edgePaddingLow.getValues<int64_t>()[i];
-      int64_t highPad = edgePaddingHigh.getValues<int64_t>()[i];
-      int64_t dim = input.getType().getShape()[i];
-      int64_t start = -std::min((int64_t)0, lowPad);
-      int64_t end = dim + std::min((int64_t)0, highPad);
-      if (index[i] < start || index[i] >= end) {
-        valid = false;
-        break;
-      }
-    }
-    if (!valid) {
-      nextIndex(index, input.getType().getShape());
-      continue;
-    }
-    uint64_t resultIdx = 0;
-    uint64_t idxMultiplyer = 1;
-    for (int64_t i = index.size() - 1; i >= 0; --i) {
-      int64_t interiorPaddingValue =
-          (interiorPadding) ? interiorPadding.getValues<int64_t>()[i] : 0;
-      resultIdx += (edgePaddingLow.getValues<int64_t>()[i] +
-                    index[i] * (interiorPaddingValue + 1)) *
-                   idxMultiplyer;
-      idxMultiplyer *= returnType.getDimSize(i);
-    }
-    auto value = input.getValues<T>()[operandIdx];
-    result[resultIdx] = value;
-    nextIndex(index, input.getType().getShape());
-  }
-  return DenseElementsAttr::get(returnType, result);
-}
-
-template <class T, class F = T(T, T)>
-DenseElementsAttr
-ReduceWindowOpFold(DenseElementsAttr inputAttr, RankedTensorType outputType,
-                   ArrayRef<int64_t> dimensions, ArrayRef<int64_t> strides,
-                   T &initValue, F &functor) {
-  auto inputShape = inputAttr.getType().getShape();
-  auto outputShape = outputType.getShape();
-  llvm::SmallVector<T> output(outputType.getNumElements(), initValue);
-  auto values = inputAttr.getValues<T>();
-  auto tensorIndexToScalarIndex = [](ArrayRef<uint64_t> index,
-                                     ArrayRef<int64_t> shape) {
-    assert(shape.size() == index.size());
-    uint64_t resultIdx = 0;
-    uint64_t idxMultiplyer = 1;
-    for (int i = index.size() - 1; i >= 0; --i) {
-      resultIdx += index[i] * idxMultiplyer;
-      idxMultiplyer *= shape[i];
-    }
-    return resultIdx;
-  };
-  auto nextIndex =
-      [](llvm::SmallVector<uint64_t> &index, llvm::ArrayRef<int64_t> windowDims,
-         llvm::ArrayRef<int64_t> windowStrides, llvm::ArrayRef<int64_t> shape) {
-        for (int64_t i = index.size() - 1; i >= 0; --i) {
-          index[i] += windowStrides[i];
-          if (static_cast<int64_t>(index[i] + windowDims[i]) < shape[i])
-            return;
-          index[i] = 0;
-        }
-      };
-  auto nextKernelIndex = [](llvm::SmallVector<uint64_t> &startIndex,
-                            llvm::ArrayRef<int64_t> windowDims,
-                            llvm::SmallVector<uint64_t> &index) {
-    for (int64_t i = index.size() - 1; i >= 0; --i) {
-      ++index[i];
-      if (static_cast<int64_t>(index[i]) < (startIndex[i] + windowDims[i]))
-        return;
-      index[i] = startIndex[i];
-    }
-  };
-  uint64_t kernelNums = 1;
-  for (auto dim : dimensions) {
-    kernelNums *= dim;
-  }
-
-  llvm::SmallVector<uint64_t> startIndex(inputShape.size(), 0);
-  for (uint64_t i = 0; i < output.size(); ++i) {
-    llvm::SmallVector<uint64_t> index = startIndex;
-    T &value = output[i];
-    for (uint64_t j = 0; j < kernelNums; ++j) {
-      uint64_t scalerIndex = tensorIndexToScalarIndex(index, inputShape);
-      value = functor(value, values[scalerIndex]);
-      nextKernelIndex(startIndex, dimensions, index);
-    }
-    nextIndex(startIndex, dimensions, strides, inputShape);
-  }
-  DenseElementsAttr outputAttr = DenseElementsAttr::get(outputType, output);
-  return outputAttr;
-}
-template <class T, class F = T(T, T)>
-DenseElementsAttr Maximum(DenseElementsAttr lhsAttr, DenseElementsAttr rhsAttr,
-                          F &functor) {
-  if (!lhsAttr || !rhsAttr)
-    return {};
-  auto lhsValues = lhsAttr.getValues<T>();
-  auto rhsValues = rhsAttr.getValues<T>();
-  assert(lhsValues.size() == rhsValues.size());
-  llvm::SmallVector<T> output;
-  output.reserve(lhsValues.size());
-  for (uint64_t i = 0; i < lhsValues.size(); ++i) {
-    output.push_back(functor(lhsValues[i], rhsValues[i]));
-  }
-  DenseElementsAttr outputAttr =
-      DenseElementsAttr::get(lhsAttr.getType(), output);
-  return outputAttr;
 }
 
 bool getFloat(Type eleType, float val, APFloat &value) {
@@ -392,181 +285,69 @@ bool getFloat(Type eleType, float val, APFloat &value) {
   return true;
 }
 
-void MhloShapeValueAnalysis::visitOperation(
-    Operation *op, ArrayRef<const ShapeValueLattice *> operands,
-    ArrayRef<ShapeValueLattice *> results) {
-  LLVM_DEBUG(llvm::dbgs() << "mhlo shape value analysis on " << *op << "\n");
-  TypeSwitch<Operation *>(op)
-      .Case<mhlo::ReduceOp>([&](Operation *op) {
-        mhlo::ReduceOp reduceOp = dyn_cast<mhlo::ReduceOp>(op);
-        auto num = op->getNumResults();
-        assert(num == 1);
-        Operation &innerOp = *reduceOp.getBody().front().begin();
-        if (!dyn_cast<mhlo::MulOp>(&innerOp)) {
-          return setAllToEntryStates(results);
-        }
-        Value input = reduceOp.getInputs()[0];
-        Value output = op->getResult(0);
-
-        auto *operand = operands[0];
-        auto *inputShapeLattice = getOrCreate<ShapeLattice>(input);
-        inputShapeLattice->useDefSubscribe(this);
-        auto *outputShapeLattice = getOrCreate<ShapeLattice>(output);
-        outputShapeLattice->useDefSubscribe(this);
-
-        if (operand->getValue().isUninitialized()) {
-          return;
-        }
-        if (inputShapeLattice->getValue().isUninitialized()) {
-          return;
-        }
-        if (outputShapeLattice->getValue().isUninitialized()) {
-          return;
-        }
-
-        if (!operand->getValue().getConstantValue()) {
-          return setAllToEntryStates(results);
-        }
-        RankedTensorType inputType =
-            dyn_cast<RankedTensorType>(inputShapeLattice->getValue().getType());
-        if (!inputType || !inputType.hasStaticShape()) {
-          return setAllToEntryStates(results);
-        }
-        RankedTensorType outputType = dyn_cast<RankedTensorType>(
-            outputShapeLattice->getValue().getType());
-        if (!outputType || !outputType.hasStaticShape()) {
-          return setAllToEntryStates(results);
-        }
-
-        auto inputShape = inputType.getShape();
-        auto dimensions = reduceOp.getDimensions().getValues<int64_t>();
-        llvm::SmallVector<int64_t> windowDimensions(inputShape.size(), 1);
-        for (auto dim : dimensions) {
-          windowDimensions[dim] = inputShape[dim];
-        }
-        llvm::SmallVector<int64_t> windowStrides(inputShape.size(), 1);
-        DenseElementsAttr inputAttr =
-            dyn_cast<DenseElementsAttr>(operand->getValue().getConstantValue());
-
-        Attribute outAttr;
-        if (isa<FloatType>(outputType.getElementType())) {
-          APFloat initValue(
-              cast<FloatType>(outputType.getElementType()).getFloatSemantics());
-          assert(getFloat(outputType.getElementType(), 1.0, initValue));
-          std::function<APFloat(APFloat, APFloat)> mulFunctor =
-              [](APFloat l, APFloat r) { return l * r; };
-          outAttr = ReduceWindowOpFold(inputAttr, outputType, windowDimensions,
-                                       windowStrides, initValue, mulFunctor);
-        } else if (isa<IntegerType>(outputType.getElementType())) {
-          APInt initValue(inputAttr.getValues<APInt>()[0].getBitWidth(), 1);
-          std::function<APInt(APInt, APInt)> mulFunctor = [](APInt l, APInt r) {
-            return l * r;
-          };
-          outAttr = ReduceWindowOpFold(inputAttr, outputType, windowDimensions,
-                                       windowStrides, initValue, mulFunctor);
-        } else {
-          return SparseConstantPropagation::visitOperation(op, operands,
-                                                           results);
-        }
-
-        auto lattice = results[0];
-        propagateIfChanged(lattice, lattice->join(mlir::dataflow::ConstantValue(
-                                        outAttr, op->getDialect())));
-      })
-      .Case<mhlo::ComputeReshapeShapeOp>([&](Operation *op) {
-        mhlo::ComputeReshapeShapeOp computeReshapeShapeOp =
-            dyn_cast<mhlo::ComputeReshapeShapeOp>(op);
-        Value dynamicShapeV = computeReshapeShapeOp.getDynamicShape();
-        auto *boundedShapeValue =
-            getOrCreate<BoundedValueLattice>(dynamicShapeV);
-        boundedShapeValue->useDefSubscribe(this);
-
-        const ShapeValueLattice *product = operands[0];
-        const ShapeValueLattice *shapeValue = operands[1];
-        if (product->getValue().isUninitialized()) {
-          return;
-        }
-        if (shapeValue->getValue().isUninitialized()) {
-          return;
-        }
-        if (boundedShapeValue->getValue().isUninitialized()) {
-          return;
-        }
-
-        if (!product->getValue().getConstantValue()) {
-          return setAllToEntryStates(results);
-        }
-        if (!shapeValue->getValue().getConstantValue() &&
-            boundedShapeValue->getValue().isUnknown()) {
-          return setAllToEntryStates(results);
-        }
-        Attribute productAttr = product->getValue().getConstantValue();
-        Attribute constShapeAttr = shapeValue->getValue().getConstantValue();
-        Attribute upperShapeAttr = boundedShapeValue->getValue().upper();
-        if (constShapeAttr) {
-          assert(!upperShapeAttr);
-        } else {
-          assert(upperShapeAttr);
-          constShapeAttr = upperShapeAttr;
-        }
-
-        Attribute resAttr = constShapeAttr;
-        ShapeValueLattice *lattice = results[0];
-        // in some cases, the shape in computeReshapeShapeOp is dense<[-1, x,
-        // ....]>, we need calculate firstly
-        do {
-          auto denseInt =
-              dyn_cast_or_null<DenseIntElementsAttr>(constShapeAttr);
-          if (denseInt == nullptr) {
-            break;
-          }
-          auto dataType = dyn_cast<IntegerType>(denseInt.getElementType());
-          // is int32
-          if (dataType == nullptr || dataType.isUnsigned() ||
-              dataType.getWidth() != 32) {
-            break;
-          }
-          llvm::SmallVector<int32_t> shape =
-              llvm::to_vector(denseInt.getValues<int32_t>());
-
-          // check whether has dimSize < 0, aka dynamic in mhlo
-          int cntDynamic = llvm::count_if(
-              shape, [](int32_t dimSize) { return dimSize < 0; });
-
-          if (cntDynamic == 1) {
-            if (auto num = dyn_cast_or_null<IntegerAttr>(productAttr)) {
-              int64_t number = num.getInt();
-              if (number < 0) {
-                break;
-              }
-
-              int32_t index = K_INITIAL;
-              for (auto elem : llvm::enumerate(shape)) {
-                if (elem.value() < 0) {
-                  index = elem.index();
-                } else {
-                  number /= elem.value();
-                }
-              }
-              assert(index != K_INITIAL);
-              shape[index] = number;
-              resAttr = DenseIntElementsAttr::get(denseInt.getType(), shape);
-            }
-          }
-        } while (0);
-
-        LLVM_DEBUG(llvm::dbgs() << "Folded to constant: " << resAttr << "\n");
-        propagateIfChanged(lattice, lattice->join(mlir::dataflow::ConstantValue(
-                                        resAttr, op->getDialect())));
-      })
-      .Default([&](Operation *op) {
-        ShapeValueAnalysis::visitOperation(op, operands, results);
-      });
+void MhloBoundedValueAnalysis::setToEntryState(BoundedValueLattice *lattice) {
+  value_analysis::BoundedValueKnowledge next =
+      value_analysis::BoundedValueKnowledge::getUnknownValue();
+  propagateIfChanged(lattice, lattice->join(next));
 }
 
 void MhloBoundedValueAnalysis::visitOperation(
     Operation *op, ArrayRef<const BoundedValueLattice *> operands,
-    // ArrayRef<ShapeLattice *> shapeLattices,
+    ArrayRef<BoundedValueLattice *> results) {
+  LLVM_DEBUG(llvm::dbgs() << "shape value analysis on " << *op << "\n");
+
+  SmallVector<ShapeLattice *> shapeLattices;
+  SmallVector<ShapeValueLattice *> shapeValueLattices;
+  ValueTypeModificatoinRAII valueTypeModification;
+  auto inputs = op->getOperands();
+  bool missingValue = false;
+
+  for (int i = 0; i < static_cast<int>(inputs.size()); ++i) {
+    auto input = inputs[i];
+    auto *boundedValueLattice = operands[i];
+    auto *shapeLattice = getOrCreate<ShapeLattice>(input);
+    shapeLattice->useDefSubscribe(this);
+    auto *shapeValueLattice = getOrCreate<ShapeValueLattice>(input);
+    shapeValueLattice->useDefSubscribe(this);
+    shapeLattices.push_back(shapeLattice);
+    shapeValueLattices.push_back(shapeValueLattice);
+
+    if (boundedValueLattice->getValue().isUninitialized()) {
+      missingValue = true;
+    }
+
+    if (shapeLattice->getValue().isUninitialized()) {
+      missingValue = true;
+    } else {
+      if (dyn_cast<ShapedType>(input.getType())) {
+        auto shapeKnowledge = shapeLattice->getValue();
+        if (shapeKnowledge && *shapeKnowledge.dtype) {
+          valueTypeModification.Push(input, shapeKnowledge.getType());
+        }
+      }
+    }
+
+    if (shapeValueLattice->getValue().isUninitialized()) {
+      missingValue = true;
+    }
+  }
+  if (missingValue)
+    return;
+  bool shapeValueReady =
+      std::all_of(shapeValueLattices.begin(), shapeValueLattices.end(),
+                  [](const ShapeValueLattice *shapeValueLattice) {
+                    return shapeValueLattice->getValue().getConstantValue();
+                  });
+  if (shapeValueReady) {
+    return setAllToEntryStates(results);
+  }
+
+  return visitOperation(op, operands, /*shapeLattices,*/
+                        shapeValueLattices, results);
+}
+
+void MhloBoundedValueAnalysis::visitOperation(
+    Operation *op, ArrayRef<const BoundedValueLattice *> operands,
     ArrayRef<ShapeValueLattice *> shapeValueLattices,
     ArrayRef<BoundedValueLattice *> results) {
   TypeSwitch<Operation *>(op)
@@ -711,7 +492,7 @@ void MhloBoundedValueAnalysis::visitOperation(
             DenseIntElementsAttr::get(type, interiorPad);
         llvm::SmallVector<int64_t> padShape;
         auto inputShape = inputType.getShape();
-        for (int i = 0; i < inputShape.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(inputShape.size()); ++i) {
           padShape.push_back(inputShape[i] + lowPad[i] + highPad[i]);
         }
         RankedTensorType padType =
@@ -942,7 +723,7 @@ void MhloBoundedValueAnalysis::visitOperation(
         auto shapeEqual = [](ArrayRef<int64_t> l, ArrayRef<int64_t> r) {
           if (l.size() != r.size())
             return false;
-          for (int i = 0; i < l.size(); ++i) {
+          for (int i = 0; i < static_cast<int>(l.size()); ++i) {
             if (l[i] != r[i])
               return false;
           }
@@ -967,7 +748,7 @@ void MhloBoundedValueAnalysis::visitOperation(
         auto *rhsShapeValueLattice = shapeValueLattices[2];
 
         mhlo::CompareOp compareOp =
-            dyn_cast<mhlo::CompareOp>(selectOp.getPred().getDefiningOp());
+            selectOp.getPred().getDefiningOp<mhlo::CompareOp>();
         if (!compareOp) {
           return setAllToEntryStates(results);
         }
@@ -1076,7 +857,7 @@ void MhloBoundedValueAnalysis::visitOperation(
         }
         SmallVector<mlir::Attribute> lowerAttrs;
         SmallVector<mlir::Attribute> upperAttrs;
-        for (int i = 0; i < inputs.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(inputs.size()); ++i) {
           auto *operand = operands[i];
           auto *shapeValueLattice = shapeValueLattices[i];
           if (operand->getValue().isUnknown() &&
@@ -1125,7 +906,7 @@ void MhloBoundedValueAnalysis::visitOperation(
         }
         SmallVector<mlir::Attribute> lowerAttrs;
         SmallVector<mlir::Attribute> upperAttrs;
-        for (int i = 0; i < inputs.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(inputs.size()); ++i) {
           auto *operand = operands[i];
           auto *shapeValueLattice = shapeValueLattices[i];
           if (operand->getValue().isUnknown() &&
@@ -1153,61 +934,6 @@ void MhloBoundedValueAnalysis::visitOperation(
         foldOp(op, lowerAttrs, upperAttrs, results);
       })
       .Default([&](Operation *op) { setAllToEntryStates(results); });
-}
-
-void MhloBoundedValueAnalysis::visitOperation(
-    Operation *op, ArrayRef<const BoundedValueLattice *> operands,
-    ArrayRef<BoundedValueLattice *> results) {
-  LLVM_DEBUG(llvm::dbgs() << "shape value analysis on " << *op << "\n");
-
-  SmallVector<ShapeLattice *> shapeLattices;
-  SmallVector<ShapeValueLattice *> shapeValueLattices;
-  ValueTypeModificatoinRAII valueTypeModification;
-  auto inputs = op->getOperands();
-  bool missingValue = false;
-
-  for (int i = 0; i < inputs.size(); ++i) {
-    auto input = inputs[i];
-    auto *boundedValueLattice = operands[i];
-    auto *shapeLattice = getOrCreate<ShapeLattice>(input);
-    shapeLattice->useDefSubscribe(this);
-    auto *shapeValueLattice = getOrCreate<ShapeValueLattice>(input);
-    shapeValueLattice->useDefSubscribe(this);
-    shapeLattices.push_back(shapeLattice);
-    shapeValueLattices.push_back(shapeValueLattice);
-
-    if (operands[i]->getValue().isUninitialized()) {
-      missingValue = true;
-    }
-
-    if (shapeLattice->getValue().isUninitialized()) {
-      missingValue = true;
-    } else {
-      if (dyn_cast<ShapedType>(input.getType())) {
-        auto shapeKnowledge = shapeLattice->getValue();
-        if (shapeKnowledge && *shapeKnowledge.dtype) {
-          valueTypeModification.Push(input, shapeKnowledge.getType());
-        }
-      }
-    }
-
-    if (shapeValueLattice->getValue().isUninitialized()) {
-      missingValue = true;
-    }
-  }
-  if (missingValue)
-    return;
-  bool shapeValueReady =
-      std::all_of(shapeValueLattices.begin(), shapeValueLattices.end(),
-                  [](const ShapeValueLattice *shapeValueLattice) {
-                    return shapeValueLattice->getValue().getConstantValue();
-                  });
-  if (shapeValueReady) {
-    return setAllToEntryStates(results);
-  }
-
-  return visitOperation(op, operands, /*shapeLattices,*/
-                        shapeValueLattices, results);
 }
 
 void MhloBoundedValueAnalysis::foldOp(Operation *op,
